jobmaster: move system level jobmaster code outside demo
