+++ master/jobmaster/system/interface.go
@@ -1,4 +1,4 @@
-package jobmaster
+package system
 
 import (
 	"context"
